requirements.txt
# 核心依赖
langchain==0.3.0,<0.4.0
langchain-core
pydantic>=2.0.0,<3.0.0

# UI和图像处理
gradio>=4.0.0,<5.0.0
pillow>=10.0.0,<11.0.0

# API和请求
requests

# 模型集成
ollama>=0.1.5,<0.2.0

# 配置和环境
pyyaml
python-dotenv>=1.0.0,<2.0.0

# 可选依赖，取消注释以启用
# fastapi>=0.100.0,<0.101.0
# uvicorn>=0.23.0,<0.24.0
